Add test for safeParseInt
