+++ package.json
@@ -32,6 +32,7 @@
     ],
     "reporter": [
       "text",
+      "html",
       "lcov"
     ]
   },
